Overflow in the serial receiver causes system to crash (#20)
If the serial receiver is enabled and the receive buffer gets full, the system crashed.
Reason: The Serial::interruptHandler() will never read out the receive buffer register RBR if readBufferFull() is true. As a result the receive interrupt is not acknowledged and the ISR is constantly called. Non ISR routines cease.