Add binaries for testing on a LPC1115. After flashing the controller enters the bootloader and waits for the main firmware to be send over the bus. To unlock the controller, either the UID is required, or when omitted in the PC tool, device needs to be set in programming mode by the button to indicate that we have valid, physical access. The PC tool will also display the UID, so that it can be used next time to unlock the device via the bus.
In order to get a working main firmware, the correct location in the
flash needs to be adjusted in the linker script.
Project Properties --> C/C++ Build --> MCU Settings, change flash
location from 0x00 to 0x3000 to ensure proper alignment in flash.

Please note, after flashing the main firmware, the bootloader will start
this one instead. Calling the bootloader from the main firmware needs to
be implemented in the main firmware by writing the magic word. See
bootloader source.

Last but not least, this is work in progress. Contributions are welcome!