Bus-Updater V0.51:
- bump version to V0.51 (0x1051)
- fix diff updater to use application base address found in boot
descriptor. Diff updater works now with different offset for main
firmware and bootloader size, as long as the main firmware offset does
not change. In case, the full update feature has to be used once, then
the diff updater should work again.

To fully fix this, the PC tool needs to send the position of the
firmware before starting the diff updater.