Add source of Java based PC update tool for the Selfbus Bus updater.
This code is based on the work from
https://github.com/baycom/selfbus-updater and back-ported from the
extensions and modifications made by https://github.com/pavkriz/knxduino

This code is still work on progress, especially the differential updater
feature might fail depending on the already loaded firmware and the new
one to be send.
In case of a failure, usually a full download works well and afterwards
the differential feature can be used for small updates.

Latest changes:
Backported knxduino differential bus updater algorithm to selfbus ARM
updater. This is the first running proof-of-concept tested with an
out8-bcu1 device.
== DEBUG output active! ==
- Various bugfixes and adoptions to get it work with selfbus devices
- changed debug output
- SLF4J with gradle to satisfy calimero (suppress warnings)

(sh 20201122)