Change implementation to use more OOP instead of build configs
The app has to be changed to instantiate an instance of the matching
class: BCU1, BCU2, MASK0701, MASK0705, SYSTEMB instead of referencing
a lib compiled with the matching build config.
Some methods like objectWrite are now moved into objects and can no
longer be called in C-style. objectWrite() has to be changed to
bcu->comObjects->objectWrite()