[sblib] Detect collisions also when sending the most-significant high bits
Previously, collision detection was only enabled when there was at least
one more 0 to send, and disabled when only 1s remained to send. For
example, for 0xFD we need to send 11 1111 1101 0. For bit 0 (least
significant bit), collision detection was enabled. But for bits 2-9
it was disabled.

Technically, the reason for this was that state SEND_BITS_OF_BYTE
transitioned directly to SEND_BIT_0 or SEND_END_OF_TX when the stop bit
transmission was included in the timer it sets up, and these states do
not detect collisions.

To remedy this, all high bits are now sent in SEND_WAIT_FOR_HIGH_BIT_END
and a new state SEND_END_OF_BYTE transitions to the correct next state.