Ensure telegrams are repeated if necessary
After sending out a telegram, we wait for an acknowledge frame.
Previously, when we did not receive an ACK but a real telegram, we did
not repeat the telegram anymore.

With this fix, receiving another unrelated telegram does not clear the
`repeatTelegram` flag anymore. Also, only the first acknowledge frame we
receive within the allotted time window is evaluated. Consecutive ACKs
or ACKs sent in response to other telegrams do not prevent our
repetition anymore.