[sblib] Fix order of telegrams in dump
As there are separate variables and buffers for TX and RX telegram
dumps, both of these can be filled at the same time. Whichever one
we opt to output first causes some of the telegrams to show up in
swapped order compared to what happened on the bus, and that is
confusing.

So output TX first if it happened first, then RX, then TX if necessary.

Consequently, also move the TXAck output after RX.