[sblib] Don't send multiple T_ACK for repeated N_DATA_INDIVIDUAL
The Data Link Layer filters out repeated telegrams, provided that the repetition follows directly after the original telegram. If another telegram sneaks in (e.g. due to higher priority), we encounter such repeated telegrams in the Transport Layer.

Per KNX Spec 2.1, Chapter 3/3/4 Section 5.4.4.3 p.27, this is Event E05 and its corresponding Action A3. The spec says to send another T_ACK for such a repeated telegram, but here's the catch:
If the client does not implement Style 3, such as calimero-core 2.5.1 at the time of this writing (2023/04/02), it will see a duplicate T_ACK in state OPEN_IDLE (events E08/E09) and consequently close the connection.

We prevent this by staying silent and intentionally disobeying the spec.