[sblib] Choose different approach for buffer handling in Style 3
Introduce dedicated buffer for connection-oriented telegrams. Ensure all
responses are sent via this buffer such that they can be repeated if
necessary. Introduce states for both buffers to make the flow easier to
understand and maintain.

This also requires to move the implementation of Master Reset PDU to
`BcuDefault::loop()` to ensure the response telegram can be repeated
if required.