[sblib] Encapsulate `TLayer4::send(Connected)Telegram`
Use a well-defined interface to hand out pointers to the correct buffer
to write telegrams to. This fixes the implementation of
`APCI_ADC_READ_PDU` (and others) that can be used in connectionless as
well as connection-oriented communication. Fix the `ADC_READ` test case
along the way, and add a new one for the connectionless variant.