[sblib] Implement Transport Layer Style 3 overlapped requests
This one is easy to miss and requires yet another send buffer in
`TLayer4`, but there is test case 20b that captures it nicely. Due to
the more relaxed handling of out-of-sequence messages it can happen that
one receives a message even before the previous one has been T_ACK'ed.

With this change, we reach and achieve 100% spec test case compliance on
Transport Layer Style 3 :-)