[sblib] Fix bug: ACK frame with wrong parity caused wrong repetition
Seen in a log yesterday: After sending a frame we saw an acknowledge
frame with `currentByte==0xC4` and `parity==false`. That caused us to
repeat the telegram as the send buffer was not cleared, but we repeated
the telegram with the repeat flag not set, i.e. we sent the original
telegram again without modification.