[sblib] Remove unnecessary duplicate register accesses
Since registers are declared `volatile`, sequences in the style of
`REG &= val1; REG |= val2;` read and write the register twice as the
compiler is forbidden to optimize these accesses. Using the sequence
`REG = (REG & val1) | val2;` instead reduces this to a single read
and a single write access.

Reduces binary size of Bootloader by 12 bytes and every invocation of
an affected function by 2 instructions.