Simplify condition
`flushUserMemory` waits for `bus->idle()`, so this conjunction is
superfluous.

`bus->telegramLen == 0` means that there is no received telegram to
process, which can only happen when a slow app does not process a
received telegram within the telegram separation time window. That in
turn means it's probably ok to defer the processing until after the
IAP operation.