Pause bus access before IAP operations and resume afterwards
This commit implements cooperative bus access for IAP erase and flash
operations.

Previously, when we triggered one of these IAP operations, the `Bus`
timer continued to run, depending on the concrete timing even with PWM
turned on. That caused all sorts of funny 0 bits on the KNX bus, and
a bunch of utterly broken (non-)telegrams.

From now on, we wait for the `Bus` class to arrive in a safe state for
such operations and then pause bus access. After the operation
completes, we resume bus access and sync to the bus as soon as possible.

Safe states are:
* Idle
* Received a telegram and sent out an acknowledge frame
* Sent a telegram completely, i.e. received an LL_ACK or ran out of
  retries
* Did not send an enqueued telegram yet (depending on situation)