Fix handling of start bits
Start bits within a frame are allowed to arrive a whopping 30us early
(KNX spec 2.1 chapter 3/2/2 section 2.3.1 p. 35,
p_class = inner_frame_char). When they do, we must not switch to RX as
collision detection must be performed on data bits only. Instead, the
purpose of the start bit is to synchronize, so that is what we need to
do.

For bits that arrive even earlier, KNX spec 2.1 chapter 3/2/2 section
1.4.1 p. 24 specifies that for p_class = start_of_Frame we must perform
bus free detection, and for p_class in (inner_Frame_char, ack_char) we
must not.