Fix implementation of Master Reset
Always send out a `APCI_MASTER_RESET_RESPONSE_PDU`, not only when
receiving a master reset request with "magic word".

Set `BOOTLOADER_MAGIC_WORD` only right before the system reset and even
disable interrupts before doing so, because this data overwrites an area
of the interrupt table and it's better to be safe than sorry.