[Updater] Use calimero default response message timeout
Using 3 seconds is too low, because when telegrams get dropped in
transport layer connections, they are only repeated after 3 seconds.
That means on every dropped telegram (e.g. due to high bus load) we saw
a KNXTimeoutException on the console and in the logs, when in fact
everything was fine.

Calimero has a default timeout of 5 seconds, which makes sense to use.