Fix timing of retransmit after collision
On collision, we switch to `INIT` state, which waits for an idle bus for
50 bit times. Then we go through `IDLE` to
`WAIT_50BT_FOR_NEXT_RX_OR_PENDING_TX_OR_IDLE`, which waits another 1 bit
time (`PRE_SEND_TIME`) before actually starting to send. Since this is
a disadvantage of 1 bit time for us, we need to account for it in the
setup of the timer in `INIT` state.

Similarly, every falling edge in `INIT` state should not reset the timer
to zero, because we already spent some time processing the edge, and the
timer should be aligned to the falling edge. Thus, set the timer to a
mostly correct value instead.