Fix condition to detect last transmission chance (bus access fairness)
One parenthesis was wrong in the original code introduced in commit
229bc36b6a69f15a56c34a869f394dbdcc43ad02. Writing the code in a better
readable way makes it way easier to understand.

This caused the last transmission to not be timed as early as possible
on a highly loaded bus, with logs showing more than four consecutive
collisions while trying to send out a `SEND_LAST_ERROR` telegram.