Limit number of TL4 connection control telegram retries
If a telegram with highest priority cannot be sent out successfully
after several tries, it does not make sense to try infinitely as that
floods the bus.

For example, we have seen some IP interface always respond with LL_BUSY
when there are too many open IP connections. It's not helpful to flood
the bus with KNX telegrams during that time, so stop doing that.