[BL] Allow `UPD_REQUEST_UID` without active programming mode
After a long discussion in RocketChat bootloader channel
https://chat.selfbus.org/channel/bootloader?msg=eh2Wp2QN4wrjjtkLc
we came to the conclusion:
"So we all agree, when it comes to attack scenarios, the UID doesn't help at all,
and can be removed if there are no other reasons"
This is the first step. Until implementation of `APP_VERSION` checks, we keep the `UID`
so accidental updating the wrong device can be avoided by calling the Updater with `--uid`.