[sblib] Ensure the timer does not trigger prematurely in `Bus::resume`
Setting the timer value to zero does not reset the Prescale Counter.
Consequently, even the next clock tick could cause the timer to trigger
already.

For the usage of `startSendingImmediately` from `sendTelegram` that was
fine. However, overwriting the timer match value in `resume` was not
reliable enough as it could sometimes simply be too late.

Fix this by using `timer.reset` instead of `timer.value(0)`, do that
before setting the new match value, and pass the new match value as
parameter instead of overwriting it later.