[sblib] Use multiples of 104us within frames and exact values outside
In recent traces, we saw that our inter-frame timings (wait times
between frames) are much better now. Unfortunately, trying to be as
exact as possible is hard within a frame, because the timing then
suddenly differs depending on the sequences of bits we need to send.

Thus, go back to integer multiples of 104us for the individual bits.