[sblib] Send T_Disconnect on Master Reset even when disconnected
Depending on the exact timing, the Updater/calimero might send a
T_Disconnect message before the device gets a chance to. Then, we did
not send an own T_Disconnect anymore, as we were already in state
CLOSED. Unfortunately, calimero logs a "negative confirmation" in such
a case.

Fix this by always sending a T_Disconnect message if there is an open
connection at the point in time when we receive the restart request.
This ensures we practically always send a T_Disconnect, even if we
received a T_Disconnect before.