[Updater] Don't fake start or end address in `toString`
This was not consistent. `toString` should return the actual values
and not `INVALID_ADDRESS`, even if the BootDescriptor is not valid. It's
still not 100% consistent because the long addresses are "cast"
to short for smaller log output.
Right now we are dealing with 16bit addresses, so this ok, but needs later
attention.