[Updater] New option `--ip-tunnel-reconnect` for some IP-Interfaces
Allows to reconnect the KNX IP tunnel on specified sequence number.
May help with some IP-Interfaces which are not fully Spec. compliant.
E.g. for Loxone Miniserver Gen 1. set it to 245.

Getting the current tunnel connection sequence number is quite hacky
reflection stuff. This is why it is handled separately in the
new subclass `DeviceManagementReflect`